+++ handler/transport/oauth2.go
@@ -8,174 +8,91 @@ import (
 	"io/ioutil"
 	"net/http"
 	"net/url"
-	"strings"
 
-	"github.com/avenga/couper/cache"
 	"github.com/avenga/couper/config"
 	"github.com/avenga/couper/config/request"
 	"github.com/avenga/couper/errors"
 	"github.com/avenga/couper/eval"
-	"github.com/avenga/couper/internal/seetie"
 )
-
-var _ http.RoundTripper = &OAuth2{}
 
 // OAuth2 represents the transport <OAuth2> object.
 type OAuth2 struct {
-	backend  http.RoundTripper
-	config   *config.OAuth2
-	memStore *cache.MemoryStore
-	next     http.RoundTripper
+	Backend http.RoundTripper
+	config  config.OAuth2
 }
 
-type OAuth2Credentials struct {
-	ClientID                string
-	ClientSecret            string
-	Scope                   *string
-	StorageKey              string
-	TokenEndpointAuthMethod *string
+type OAuth2RequestConfig struct {
+	Code         *string
+	CodeVerifier *string
+	RedirectURI  *string
+	StorageKey   string
 }
 
-// NewOAuth2 creates a new <http.RoundTripper> object.
-func NewOAuth2(conf *config.OAuth2, memStore *cache.MemoryStore,
-	backend, next http.RoundTripper) (http.RoundTripper, error) {
-	const grantType = "client_credentials"
-	if conf.GrantType != grantType {
-		return nil, errors.Backend.Label(conf.BackendName).Message("grant_type not supported: " + conf.GrantType)
-	}
-
-	return &OAuth2{
-		backend:  backend,
-		config:   conf,
-		memStore: memStore,
-		next:     next,
-	}, nil
-}
-
-// RoundTrip implements the <http.RoundTripper> interface.
-func (oa *OAuth2) RoundTrip(req *http.Request) (*http.Response, error) {
-	credentials, err := oa.getCredentials(req)
-	if err != nil {
-		return nil, errors.Backend.Label(oa.config.BackendName).With(err)
-	}
-
-	if data := oa.memStore.Get(credentials.StorageKey); data != "" {
-		token, terr := oa.readAccessToken(data)
-		if terr != nil {
-			return nil, errors.Backend.Label(oa.config.BackendName).Message("token read error").With(terr)
-		}
-
-		req.Header.Set("Authorization", "Bearer "+token)
-
-		return oa.next.RoundTrip(req)
-	}
-
-	tokenReq, err := oa.newTokenRequest(req.Context(), credentials)
-	if err != nil {
-		return nil, err
-	}
-
-	tokenRes, err := oa.backend.RoundTrip(tokenReq)
-	if err != nil {
-		return nil, err
-	}
-
-	if tokenRes.StatusCode != http.StatusOK {
-		return nil, errors.Backend.Label(oa.config.BackendName).Message("token request failed")
-	}
-
-	tokenResBytes, err := ioutil.ReadAll(tokenRes.Body)
-	if err != nil {
-		return nil, errors.Backend.Label(oa.config.BackendName).Message("token request read error").With(err)
-	}
-
-	token, err := oa.updateAccessToken(string(tokenResBytes), credentials.StorageKey)
-	if err != nil {
-		return nil, errors.Backend.Label(oa.config.BackendName).Message("token update error").With(err)
-	}
-
-	req.Header.Set("Authorization", "Bearer "+token)
-
-	res, err := oa.next.RoundTrip(req)
-
-	if res != nil && res.StatusCode == http.StatusUnauthorized {
-		oa.memStore.Del(credentials.StorageKey)
-
-		ctx := req.Context()
-		if retries, ok := ctx.Value(request.TokenRequestRetries).(uint8); !ok || retries < *oa.config.Retries {
-			ctx = context.WithValue(ctx, request.TokenRequestRetries, retries+1)
-
-			req.Header.Del("Authorization")
-			*req = *req.WithContext(ctx)
-
-			return oa.RoundTrip(req)
-		}
-	}
-
-	return res, err
-}
-
-func (oa *OAuth2) getCredentials(req *http.Request) (*OAuth2Credentials, error) {
-	content, _, diags := oa.config.Remain.PartialContent(oa.config.Schema(true))
-	if diags.HasErrors() {
-		return nil, diags
-	}
-
-	evalContext, _ := req.Context().Value(eval.ContextType).(*eval.Context)
-
-	id, idOK := content.Attributes["client_id"]
-	idv, _ := id.Expr.Value(evalContext.HCLContext())
-	clientID := seetie.ValueToString(idv)
-
-	secret, secretOK := content.Attributes["client_secret"]
-	secretv, _ := secret.Expr.Value(evalContext.HCLContext())
-	clientSecret := seetie.ValueToString(secretv)
-
-	if !idOK || !secretOK {
-		return nil, fmt.Errorf("missing credentials")
-	}
-
-	var scope, teAuthMethod *string
-
-	if v, ok := content.Attributes["scope"]; ok {
-		ctyVal, _ := v.Expr.Value(evalContext.HCLContext())
-		strVal := strings.TrimSpace(seetie.ValueToString(ctyVal))
-		scope = &strVal
-	}
-
-	if v, ok := content.Attributes["token_endpoint_auth_method"]; ok {
-		ctyVal, _ := v.Expr.Value(evalContext.HCLContext())
-		strVal := strings.TrimSpace(seetie.ValueToString(ctyVal))
-		teAuthMethod = &strVal
-	}
-
-	if teAuthMethod != nil {
+// NewOAuth2 creates a new <OAuth2> object.
+func NewOAuth2(conf config.OAuth2, backend http.RoundTripper) (*OAuth2, error) {
+	if teAuthMethod := conf.GetTokenEndpointAuthMethod(); teAuthMethod != nil {
 		if *teAuthMethod != "client_secret_basic" && *teAuthMethod != "client_secret_post" {
 			return nil, fmt.Errorf("unsupported 'token_endpoint_auth_method': %s", *teAuthMethod)
 		}
 	}
-
-	return &OAuth2Credentials{
-		ClientID:     clientID,
-		ClientSecret: clientSecret,
-		Scope:        scope,
-		// Backend is build up via config and token_endpoint will configure the backend,
-		// use the backend memory location here.
-		StorageKey:              fmt.Sprintf("%p|%s|%s", &oa.backend, clientID, clientSecret),
-		TokenEndpointAuthMethod: teAuthMethod,
+	return &OAuth2{
+		Backend: backend,
+		config:  conf,
 	}, nil
 }
 
-func (oa *OAuth2) newTokenRequest(ctx context.Context, creds *OAuth2Credentials) (*http.Request, error) {
-	post := url.Values{}
-	post.Set("grant_type", oa.config.GrantType)
+func (oa *OAuth2) GetRequestConfig(req *http.Request) (*OAuth2RequestConfig, error) {
+	return &OAuth2RequestConfig{
+		// Backend is build up via config and token_endpoint will configure the backend,
+		// use the backend memory location here.
+		StorageKey: fmt.Sprintf("%p|%s|%s", &oa.Backend, oa.config.GetClientID(), oa.config.GetClientSecret()),
+	}, nil
+}
 
-	if creds.Scope != nil {
-		post.Set("scope", *creds.Scope)
+func (oa *OAuth2) RequestToken(ctx context.Context, requestConfig *OAuth2RequestConfig) ([]byte, error) {
+	tokenReq, err := oa.newTokenRequest(ctx, requestConfig)
+	if err != nil {
+		return nil, err
 	}
-	if creds.TokenEndpointAuthMethod != nil && *creds.TokenEndpointAuthMethod == "client_secret_post" {
-		post.Set("client_id", creds.ClientID)
-		post.Set("client_secret", creds.ClientSecret)
+
+	tokenRes, err := oa.Backend.RoundTrip(tokenReq)
+	if err != nil {
+		return nil, err
+	}
+
+	tokenResBytes, err := ioutil.ReadAll(tokenRes.Body)
+	if err != nil {
+		return nil, errors.Backend.Label(oa.config.Reference()).Message("token request read error").With(err)
+	}
+
+	if tokenRes.StatusCode != http.StatusOK {
+		return nil, errors.Backend.Label(oa.config.Reference()).Messagef("token request failed, response=%q", string(tokenResBytes))
+	}
+
+	return tokenResBytes, nil
+}
+
+func (oa *OAuth2) newTokenRequest(ctx context.Context, requestConfig *OAuth2RequestConfig) (*http.Request, error) {
+	post := url.Values{}
+	grantType := oa.config.GetGrantType()
+	post.Set("grant_type", grantType)
+
+	if scope := oa.config.GetScope(); scope != nil && grantType != "authorization_code" {
+		post.Set("scope", *scope)
+	}
+	if requestConfig.RedirectURI != nil {
+		post.Set("redirect_uri", *requestConfig.RedirectURI)
+	}
+	if requestConfig.Code != nil {
+		post.Set("code", *requestConfig.Code)
+	}
+	if requestConfig.CodeVerifier != nil {
+		post.Set("code_verifier", *requestConfig.CodeVerifier)
+	}
+	teAuthMethod := oa.config.GetTokenEndpointAuthMethod()
+	if teAuthMethod != nil && *teAuthMethod == "client_secret_post" {
+		post.Set("client_id", oa.config.GetClientID())
+		post.Set("client_secret", oa.config.GetClientSecret())
 	}
 
 	// url will be configured via backend roundtrip
@@ -188,67 +105,35 @@ func (oa *OAuth2) newTokenRequest(ctx context.Context, creds *OAuth2Credentials)
 
 	outreq.Header.Set("Content-Type", "application/x-www-form-urlencoded")
 
-	if creds.TokenEndpointAuthMethod == nil || *creds.TokenEndpointAuthMethod == "client_secret_basic" {
-		auth := base64.StdEncoding.EncodeToString([]byte(creds.ClientID + ":" + creds.ClientSecret))
+	if teAuthMethod == nil || *teAuthMethod == "client_secret_basic" {
+		auth := base64.StdEncoding.EncodeToString([]byte(oa.config.GetClientID() + ":" + oa.config.GetClientSecret()))
 
 		outreq.Header.Set("Authorization", "Basic "+auth)
 	}
 
 	outCtx := context.WithValue(ctx, request.TokenRequest, "oauth2")
 
-	url, err := eval.GetContextAttribute(oa.config.Remain, outCtx, "token_endpoint")
-	if err != nil {
-		return nil, err
-	}
-
-	if url != "" {
-		outCtx = context.WithValue(outCtx, request.URLAttribute, url)
+	if tokenURL := oa.config.GetTokenEndpoint(); tokenURL != "" {
+		outCtx = context.WithValue(outCtx, request.URLAttribute, tokenURL)
 	}
 
 	return outreq.WithContext(outCtx), nil
 }
 
-func (oa *OAuth2) readAccessToken(data string) (string, error) {
+func ParseAccessToken(jsonBytes []byte) (map[string]interface{}, string, error) {
 	var jData map[string]interface{}
 
-	err := json.Unmarshal([]byte(data), &jData)
+	err := json.Unmarshal(jsonBytes, &jData)
 	if err != nil {
-		return "", err
+		return jData, "", err
 	}
 
 	var token string
 	if t, ok := jData["access_token"].(string); ok {
 		token = t
 	} else {
-		return "", fmt.Errorf("missing access token")
+		return jData, "", fmt.Errorf("missing access token")
 	}
 
-	return token, nil
-}
-
-func (oa *OAuth2) updateAccessToken(jsonString, key string) (string, error) {
-	var jData map[string]interface{}
-
-	err := json.Unmarshal([]byte(jsonString), &jData)
-	if err != nil {
-		return "", err
-	}
-
-	var token string
-	if t, ok := jData["access_token"].(string); ok {
-		token = t
-	} else {
-		return "", fmt.Errorf("missing access token")
-	}
-
-	if oa.memStore != nil {
-		var ttl int64
-		if t, ok := jData["expires_in"].(float64); ok {
-			ttl = (int64)(t * 0.9)
-		}
-
-		oa.memStore.Set(key, jsonString, ttl)
-	}
-
-	return token, nil
+	return jData, token, nil
 }