+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 Unreleased changes are available as `avenga/couper:edge` container.
 
+* **Added**
+  * OAuth2 Authorization Code Grant Flow: `beta_oauth2 {}` block; `beta_oauth_authorization_url()`, `beta_oauth_code_verifier()` and `beta_oauth_csrf_token()` functions ([#247](https://github.com/avenga/couper/pull/247))
 ---
 
 ## [1.3.1](https://github.com/avenga/couper/compare/1.3...1.3.1)